refactor: Finally pickling works properly from code blocks
Loading a pickle file was succesfull only when setting, :results output, flag for
the code block. Otherwise the file failed to load giving a crypting error in the
session interpreter about name '_' not existing. Though strangely it would load
directly from the session interpreter. Also, unpacking the data variable has to be
done on a separate code block of the same session, otherwise its as if data variable
never loaded...weird.